requirements.txt
matplotlib
fpdf
python-barcode
Pillow
